fix: add deployment RBAC validation, fix keypair policy lookup, and move resource-group CLI to admin
- Add SingleEntityActionRBACValidator to deployment get/update/destroy
  actions so non-owner users can no longer modify other users' deployments
- Fix keypair resource policy lookup using wrong column (KeyPairRow.user_id
  is email, not UUID); use KeyPairRow.user instead
- Move all resource-group CLI commands from user-facing to admin group
  since every REST endpoint requires superadmin
- Unify ResourceGroupDetailNode as the single response DTO, removing the
  incomplete ResourceGroupNode with placeholder values
- Change deployment EntityType from DEPLOYMENT to MODEL_DEPLOYMENT to
  match the RBAC element type definition
