+++ src/ai/backend/common/dto/manager/v2/resource_group/response.py
@@ -82,13 +82,13 @@ class ResourceGroupNode(BaseResponseModel):
 class CreateResourceGroupPayload(BaseResponseModel):
     """Payload for resource group creation mutation result."""
 
-    resource_group: ResourceGroupNode = Field(description="Created resource group.")
+    resource_group: ResourceGroupDetailNode = Field(description="Created resource group.")
 
 
 class UpdateResourceGroupPayload(BaseResponseModel):
     """Payload for resource group update mutation result."""
 
-    resource_group: ResourceGroupNode = Field(description="Updated resource group.")
+    resource_group: ResourceGroupDetailNode = Field(description="Updated resource group.")
 
 
 class DeleteResourceGroupPayload(BaseResponseModel):
