+++ src/ai/backend/manager/services/deployment/actions/base.py
@@ -3,17 +3,37 @@ from typing import override
 from ai.backend.common.data.permission.types import EntityType
 from ai.backend.manager.actions.action import BaseAction
 from ai.backend.manager.actions.action.scope import BaseScopeAction
+from ai.backend.manager.actions.action.single_entity import (
+    BaseSingleEntityAction,
+    BaseSingleEntityActionResult,
+)
+from ai.backend.manager.actions.action.types import FieldData
 
 
 class DeploymentBaseAction(BaseAction):
     @override
     @classmethod
     def entity_type(cls) -> EntityType:
-        return EntityType.DEPLOYMENT
+        return EntityType.MODEL_DEPLOYMENT
 
 
 class DeploymentScopeAction(BaseScopeAction):
     @override
     @classmethod
     def entity_type(cls) -> EntityType:
-        return EntityType.DEPLOYMENT
+        return EntityType.MODEL_DEPLOYMENT
+
+
+class DeploymentSingleEntityAction(BaseSingleEntityAction):
+    @override
+    @classmethod
+    def entity_type(cls) -> EntityType:
+        return EntityType.MODEL_DEPLOYMENT
+
+    @override
+    def field_data(self) -> FieldData | None:
+        return None
+
+
+class DeploymentSingleEntityActionResult(BaseSingleEntityActionResult):
+    pass
