+++ src/ai/backend/manager/services/deployment/actions/destroy_deployment.py
@@ -4,20 +4,28 @@ import uuid
 from dataclasses import dataclass
 from typing import override
 
-from ai.backend.manager.actions.action import BaseActionResult
+from ai.backend.common.data.permission.types import RBACElementType
 from ai.backend.manager.actions.types import ActionOperationType
-from ai.backend.manager.services.deployment.actions.base import DeploymentBaseAction
+from ai.backend.manager.data.permission.types import RBACElementRef
+from ai.backend.manager.services.deployment.actions.base import (
+    DeploymentSingleEntityAction,
+    DeploymentSingleEntityActionResult,
+)
 
 
 @dataclass
-class DestroyDeploymentAction(DeploymentBaseAction):
+class DestroyDeploymentAction(DeploymentSingleEntityAction):
     """Action to destroy an existing deployment."""
 
     endpoint_id: uuid.UUID
 
     @override
-    def entity_id(self) -> str | None:
+    def target_entity_id(self) -> str:
         return str(self.endpoint_id)
+
+    @override
+    def target_element(self) -> RBACElementRef:
+        return RBACElementRef(RBACElementType.MODEL_DEPLOYMENT, str(self.endpoint_id))
 
     @override
     @classmethod
@@ -26,9 +34,9 @@ class DestroyDeploymentAction(DeploymentBaseAction):
 
 
 @dataclass
-class DestroyDeploymentActionResult(BaseActionResult):
+class DestroyDeploymentActionResult(DeploymentSingleEntityActionResult):
     success: bool
 
     @override
-    def entity_id(self) -> str | None:
-        return None
+    def target_entity_id(self) -> str:
+        return ""
