+++ src/ai/backend/manager/services/deployment/actions/get_deployment_by_id.py
@@ -2,19 +2,27 @@ from dataclasses import dataclass
 from typing import override
 from uuid import UUID
 
-from ai.backend.manager.actions.action import BaseActionResult
+from ai.backend.common.data.permission.types import RBACElementType
 from ai.backend.manager.actions.types import ActionOperationType
 from ai.backend.manager.data.deployment.types import ModelDeploymentData
-from ai.backend.manager.services.deployment.actions.base import DeploymentBaseAction
+from ai.backend.manager.data.permission.types import RBACElementRef
+from ai.backend.manager.services.deployment.actions.base import (
+    DeploymentSingleEntityAction,
+    DeploymentSingleEntityActionResult,
+)
 
 
 @dataclass
-class GetDeploymentByIdAction(DeploymentBaseAction):
+class GetDeploymentByIdAction(DeploymentSingleEntityAction):
     deployment_id: UUID
 
     @override
-    def entity_id(self) -> str | None:
+    def target_entity_id(self) -> str:
         return str(self.deployment_id)
+
+    @override
+    def target_element(self) -> RBACElementRef:
+        return RBACElementRef(RBACElementType.MODEL_DEPLOYMENT, str(self.deployment_id))
 
     @override
     @classmethod
@@ -23,9 +31,9 @@ class GetDeploymentByIdAction(DeploymentBaseAction):
 
 
 @dataclass
-class GetDeploymentByIdActionResult(BaseActionResult):
+class GetDeploymentByIdActionResult(DeploymentSingleEntityActionResult):
     data: ModelDeploymentData
 
     @override
-    def entity_id(self) -> str | None:
+    def target_entity_id(self) -> str:
         return str(self.data.id)
