+++ src/ai/backend/manager/services/deployment/actions/update_deployment.py
@@ -1,23 +1,31 @@
 from dataclasses import dataclass
 from typing import override
 
-from ai.backend.manager.actions.action import BaseActionResult
+from ai.backend.common.data.permission.types import RBACElementType
 from ai.backend.manager.actions.types import ActionOperationType
 from ai.backend.manager.data.deployment.types import ModelDeploymentData
+from ai.backend.manager.data.permission.types import RBACElementRef
 from ai.backend.manager.models.endpoint import EndpointRow
 from ai.backend.manager.repositories.base import Updater
-from ai.backend.manager.services.deployment.actions.base import DeploymentBaseAction
+from ai.backend.manager.services.deployment.actions.base import (
+    DeploymentSingleEntityAction,
+    DeploymentSingleEntityActionResult,
+)
 
 
 @dataclass
-class UpdateDeploymentAction(DeploymentBaseAction):
+class UpdateDeploymentAction(DeploymentSingleEntityAction):
     """Action to update an existing deployment."""
 
     updater: Updater[EndpointRow]
 
     @override
-    def entity_id(self) -> str | None:
+    def target_entity_id(self) -> str:
         return str(self.updater.pk_value)
+
+    @override
+    def target_element(self) -> RBACElementRef:
+        return RBACElementRef(RBACElementType.MODEL_DEPLOYMENT, str(self.updater.pk_value))
 
     @override
     @classmethod
@@ -26,9 +34,9 @@ class UpdateDeploymentAction(DeploymentBaseAction):
 
 
 @dataclass
-class UpdateDeploymentActionResult(BaseActionResult):
+class UpdateDeploymentActionResult(DeploymentSingleEntityActionResult):
     data: ModelDeploymentData
 
     @override
-    def entity_id(self) -> str | None:
+    def target_entity_id(self) -> str:
         return str(self.data.id)
