+++ src/ai/backend/manager/services/deployment/processors.py
@@ -6,6 +6,7 @@ from typing import TYPE_CHECKING, cast, override
 
 from ai.backend.manager.actions.monitors.monitor import ActionMonitor
 from ai.backend.manager.actions.processor import ActionProcessor
+from ai.backend.manager.actions.processor.single_entity import SingleEntityActionProcessor
 from ai.backend.manager.actions.types import AbstractProcessorPackage, ActionSpec
 from ai.backend.manager.actions.validator.base import ActionValidator
 from ai.backend.manager.actions.validators import ActionValidators
@@ -140,13 +141,19 @@ class DeploymentProcessors(AbstractProcessorPackage):
     create_legacy_deployment: ActionProcessor[
         CreateLegacyDeploymentAction, CreateLegacyDeploymentActionResult
     ]
-    update_deployment: ActionProcessor[UpdateDeploymentAction, UpdateDeploymentActionResult]
-    destroy_deployment: ActionProcessor[DestroyDeploymentAction, DestroyDeploymentActionResult]
+    update_deployment: SingleEntityActionProcessor[
+        UpdateDeploymentAction, UpdateDeploymentActionResult
+    ]
+    destroy_deployment: SingleEntityActionProcessor[
+        DestroyDeploymentAction, DestroyDeploymentActionResult
+    ]
     search_deployments: ActionProcessor[SearchDeploymentsAction, SearchDeploymentsActionResult]
     search_deployments_in_project: ActionProcessor[
         SearchDeploymentsInProjectAction, SearchDeploymentsInProjectActionResult
     ]
-    get_deployment_by_id: ActionProcessor[GetDeploymentByIdAction, GetDeploymentByIdActionResult]
+    get_deployment_by_id: SingleEntityActionProcessor[
+        GetDeploymentByIdAction, GetDeploymentByIdActionResult
+    ]
     get_deployment_policy: ActionProcessor[
         GetDeploymentPolicyAction, GetDeploymentPolicyActionResult
     ]
@@ -211,19 +218,26 @@ class DeploymentProcessors(AbstractProcessorPackage):
         validators: ActionValidators,
     ) -> None:
         # Deployment CRUD
+        rbac_single_entity_validators = [validators.rbac.single_entity]
         self.create_deployment = ActionProcessor(service.create_deployment, action_monitors)
         self.create_legacy_deployment = ActionProcessor(
             service.create_legacy_deployment, action_monitors
         )
-        self.update_deployment = ActionProcessor(service.update_deployment, action_monitors)
-        self.destroy_deployment = ActionProcessor(service.destroy_deployment, action_monitors)
+        self.update_deployment = SingleEntityActionProcessor(
+            service.update_deployment, action_monitors, validators=rbac_single_entity_validators
+        )
+        self.destroy_deployment = SingleEntityActionProcessor(
+            service.destroy_deployment, action_monitors, validators=rbac_single_entity_validators
+        )
         self.search_deployments = ActionProcessor(service.search_deployments, action_monitors)
         self.search_deployments_in_project = ActionProcessor(
             service.search_deployments_in_project,
             action_monitors,
             validators=[cast(ActionValidator, validators.rbac.scope)],
         )
-        self.get_deployment_by_id = ActionProcessor(service.get_deployment_by_id, action_monitors)
+        self.get_deployment_by_id = SingleEntityActionProcessor(
+            service.get_deployment_by_id, action_monitors, validators=rbac_single_entity_validators
+        )
         self.get_deployment_policy = ActionProcessor(service.get_deployment_policy, action_monitors)
         self.search_deployment_policies = ActionProcessor(
             service.search_deployment_policies, action_monitors
